Refactor Theme Compiler Integration Test
