+++ tests/integration/Storefront/Theme/ThemeCompilerDirectUsageTest.php
@@ -2,7 +2,6 @@
 
 namespace Shopware\Tests\Integration\Storefront\Theme;
 
-use Doctrine\DBAL\Exception;
 use League\Flysystem\Filesystem;
 use League\Flysystem\InMemory\InMemoryFilesystemAdapter;
 use PHPUnit\Framework\Attributes\CoversClass;
@@ -11,37 +10,21 @@ use Psr\Log\LoggerInterface;
 use Shopware\Core\DevOps\Environment\EnvironmentHelper;
 use Shopware\Core\Framework\Adapter\Cache\CacheInvalidator;
 use Shopware\Core\Framework\Adapter\Filesystem\Plugin\CopyBatchInputFactory;
-use Shopware\Core\Framework\App\ActiveAppsLoader;
-use Shopware\Core\Framework\App\Source\SourceResolver;
 use Shopware\Core\Framework\Context;
 use Shopware\Core\Framework\Feature;
-use Shopware\Core\Framework\Plugin;
 use Shopware\Core\Framework\Test\TestCaseBase\DatabaseTransactionBehaviour;
 use Shopware\Core\Framework\Test\TestCaseBase\KernelTestBehaviour;
-use Shopware\Core\Kernel;
-use Shopware\Core\System\SystemConfig\Service\AppConfigReader;
-use Shopware\Core\System\SystemConfig\Service\ConfigurationService;
-use Shopware\Core\System\SystemConfig\SystemConfigService;
-use Shopware\Core\System\SystemConfig\Util\ConfigReader;
-use Shopware\Core\Test\AppSystemTestBehaviour;
-use Shopware\Core\Test\TestDefaults;
-use Shopware\Storefront\Event\ThemeCompilerConcatenatedStylesEvent;
 use Shopware\Storefront\Framework\Twig\Components\TwigComponent;
+use Shopware\Storefront\Framework\Twig\Components\TwigComponentCollection;
 use Shopware\Storefront\Framework\Twig\Components\TwigComponentHelper;
-use Shopware\Storefront\Theme\Event\ThemeCompilerEnrichScssVariablesEvent;
+use Shopware\Storefront\Theme\CompilerConfiguration;
 use Shopware\Storefront\Theme\MD5ThemePathBuilder;
 use Shopware\Storefront\Theme\ScssPhpCompiler;
 use Shopware\Storefront\Theme\StorefrontPluginConfiguration\StorefrontPluginConfiguration;
 use Shopware\Storefront\Theme\StorefrontPluginConfiguration\StorefrontPluginConfigurationCollection;
-use Shopware\Storefront\Theme\StorefrontPluginConfiguration\StorefrontPluginConfigurationFactory;
-use Shopware\Storefront\Theme\StorefrontPluginRegistry;
-use Shopware\Storefront\Theme\Subscriber\ThemeCompilerEnrichScssVarSubscriber;
 use Shopware\Storefront\Theme\ThemeCompiler;
 use Shopware\Storefront\Theme\ThemeFileResolver;
 use Shopware\Storefront\Theme\ThemeFilesystemResolver;
-use Shopware\Tests\Integration\Storefront\Theme\fixtures\MockThemeCompilerConcatenatedSubscriber;
-use Shopware\Tests\Integration\Storefront\Theme\fixtures\MockThemeVariablesSubscriber;
-use Shopware\Tests\Integration\Storefront\Theme\fixtures\SimplePlugin\SimplePlugin;
 use Symfony\Component\Asset\UrlPackage;
 use Symfony\Component\Asset\VersionStrategy\EmptyVersionStrategy;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
@@ -50,9 +33,8 @@ use Symfony\Component\EventDispatcher\EventDispatcherInterface;
  * @internal
  */
 #[CoversClass(ThemeCompiler::class)]
-class ThemeCompilerTest extends TestCase
+class ThemeCompilerDirectUsageTest extends TestCase
 {
-    use AppSystemTestBehaviour;
     use DatabaseTransactionBehaviour;
     use KernelTestBehaviour;
 
@@ -75,32 +57,23 @@ class ThemeCompilerTest extends TestCase
         $this->mockSalesChannelId = '98432def39fc4624b33213a56b8c944d';
         $this->eventDispatcher = static::getContainer()->get('event_dispatcher');
 
-        /** @var TwigComponentHelper $twigComponentHelper */
-        $twigComponentHelper = static::getContainer()->get(TwigComponentHelper::class);
-
         $this->themeCompiler = new ThemeCompiler(
             $this->filesystem,
             $this->tempFilesystem,
             new CopyBatchInputFactory(),
             static::getContainer()->get(ThemeFileResolver::class),
-            $twigComponentHelper,
-            true, // debug mode
+            static::getContainer()->get(TwigComponentHelper::class),
+            true,
             $this->eventDispatcher,
             static::getContainer()->get(ThemeFilesystemResolver::class),
             ['theme' => new UrlPackage(['http://localhost'], new EmptyVersionStrategy())],
             static::getContainer()->get(CacheInvalidator::class),
             $this->createMock(LoggerInterface::class),
             new MD5ThemePathBuilder(),
-            static::getContainer()->get(ScssPhpCompiler::class), // Real SCSS compiler
+            static::getContainer()->get(ScssPhpCompiler::class),
             [],
             false
         );
-    }
-
-    protected function tearDown(): void
-    {
-        static::getContainer()->get(SourceResolver::class)->reset();
-        static::getContainer()->get(ActiveAppsLoader::class)->reset();
     }
 
     // ===================================
@@ -112,7 +85,7 @@ class ThemeCompilerTest extends TestCase
         $scssInput = '$primary-color: #ff0000; .test { color: $primary-color; }';
 
         $result = static::getContainer()->get(ScssPhpCompiler::class)->compileString(
-            new \Shopware\Storefront\Theme\CompilerConfiguration([]),
+            new CompilerConfiguration([]),
             $scssInput
         );
 
@@ -174,245 +147,12 @@ class ThemeCompilerTest extends TestCase
         $fullScss = $variables . "\n" . $testScss;
 
         $result = static::getContainer()->get(ScssPhpCompiler::class)->compileString(
-            new \Shopware\Storefront\Theme\CompilerConfiguration([]),
+            new CompilerConfiguration([]),
             $fullScss
         );
 
         static::assertStringContainsString('.button', $result);
         static::assertStringContainsString('#123456', $result);
-    }
-
-    // ===================================
-    // Event Subscriber Integration Tests
-    // ===================================
-
-    public function testEventSubscriberCanEnrichScssVariables(): void
-    {
-        $subscriber = new MockThemeVariablesSubscriber(
-            static::getContainer()->get(SystemConfigService::class)
-        );
-
-        $variables = [
-            'sw-color-brand-primary' => '#008490',
-        ];
-
-        $event = new ThemeCompilerEnrichScssVariablesEvent(
-            $variables,
-            $this->mockSalesChannelId,
-            Context::createDefaultContext()
-        );
-
-        $subscriber->onAddVariables($event);
-
-        $result = $event->getVariables();
-
-        static::assertArrayHasKey('mock-variable-black', $result);
-        static::assertSame('#000000', $result['mock-variable-black']);
-        static::assertArrayHasKey('mock-variable-special', $result);
-        static::assertSame('\'Special value with quotes\'', $result['mock-variable-special']);
-    }
-
-    public function testEventSubscriberCanModifyConcatenatedStyles(): void
-    {
-        $subscriber = new MockThemeCompilerConcatenatedSubscriber();
-
-        $styles = 'body { margin: 0; }';
-
-        $event = new ThemeCompilerConcatenatedStylesEvent($styles, $this->mockSalesChannelId);
-        $subscriber->onGetConcatenatedStyles($event);
-
-        $result = $event->getConcatenatedStyles();
-
-        static::assertStringContainsString('body { margin: 0; }', $result);
-        static::assertStringContainsString(MockThemeCompilerConcatenatedSubscriber::STYLES_CONCAT, $result);
-    }
-
-    public function testVariableEnrichmentEventAffectsCompiledOutput(): void
-    {
-        // Add a subscriber that enriches variables
-        $subscriber = new MockThemeVariablesSubscriber(
-            static::getContainer()->get(SystemConfigService::class)
-        );
-        $this->eventDispatcher->addSubscriber($subscriber);
-
-        $config = new StorefrontPluginConfiguration('TestTheme');
-        $config->setThemeConfig([
-            'fields' => [
-                'sw-color-brand-primary' => [
-                    'name' => 'sw-color-brand-primary',
-                    'type' => 'color',
-                    'value' => '#008490',
-                ],
-            ],
-        ]);
-
-        try {
-            $this->themeCompiler->compileTheme(
-                $this->mockSalesChannelId,
-                'test-theme-id',
-                $config,
-                new StorefrontPluginConfigurationCollection(),
-                false,
-                Context::createDefaultContext()
-            );
-
-            // Check that enriched variables were written
-            $variablesContent = $this->tempFilesystem->read('theme-variables.scss');
-            static::assertStringContainsString('$mock-variable-black: #000000', $variablesContent);
-        } finally {
-            $this->eventDispatcher->removeSubscriber($subscriber);
-        }
-    }
-
-    // ===================================
-    // Plugin Configuration Integration Tests
-    // ===================================
-
-    public function testCompilesWithPluginScssVariables(): void
-    {
-        $testScss = <<<'SCSS'
-.test-selector-plugin {
-    background: $simple-plugin-backgroundcolor;
-    color: $simple-plugin-fontcolor;
-}
-SCSS;
-
-        $configService = $this->getConfigurationService([
-            new SimplePlugin(true, __DIR__ . '/fixtures/SimplePlugin'),
-        ]);
-
-        $storefrontPluginRegistry = $this->getStorefrontPluginRegistry([
-            new SimplePlugin(true, __DIR__ . '/fixtures/SimplePlugin'),
-        ]);
-
-        $subscriber = new ThemeCompilerEnrichScssVarSubscriber($configService, $storefrontPluginRegistry);
-        $this->eventDispatcher->addSubscriber($subscriber);
-
-        $sysConfigService = static::getContainer()->get(SystemConfigService::class);
-        $sysConfigService->set('SimplePlugin.config.simplePluginBackgroundcolor', '#ffffff');
-        $sysConfigService->set('SimplePlugin.config.simplePluginFontcolor', '#000000');
-
-        try {
-            $result = static::getContainer()->get(ScssPhpCompiler::class)->compileString(
-                new \Shopware\Storefront\Theme\CompilerConfiguration([]),
-                '$simple-plugin-backgroundcolor: #ffffff; $simple-plugin-fontcolor: #000000; ' . $testScss
-            );
-
-            static::assertStringContainsString('.test-selector-plugin', $result);
-            static::assertStringContainsString('#ffffff', $result);
-            static::assertStringContainsString('#000000', $result);
-        } finally {
-            $this->eventDispatcher->removeSubscriber($subscriber);
-        }
-    }
-
-    public function testCompilesWithAppScssVariables(): void
-    {
-        $this->loadAppsFromDir(__DIR__ . '/fixtures/Apps/noThemeCustomCss');
-
-        $testScss = <<<'SCSS'
-.test-selector-app {
-    background: $no-theme-custom-css-backgroundcolor;
-    color: $no-theme-custom-css-fontcolor;
-}
-SCSS;
-
-        $configService = $this->getConfigurationService([
-            new SimplePlugin(true, __DIR__ . '/fixtures/SimplePlugin'),
-        ]);
-
-        $storefrontPluginRegistry = $this->getStorefrontPluginRegistry([
-            new SimplePlugin(true, __DIR__ . '/fixtures/SimplePlugin'),
-        ]);
-
-        $subscriber = new ThemeCompilerEnrichScssVarSubscriber($configService, $storefrontPluginRegistry);
-        $this->eventDispatcher->addSubscriber($subscriber);
-
-        $sysConfigService = static::getContainer()->get(SystemConfigService::class);
-        $sysConfigService->set('SwagNoThemeCustomCss.config.noThemeCustomCssBackGroundcolor', '#aabbcc');
-        $sysConfigService->set('SwagNoThemeCustomCss.config.noThemeCustomCssFontcolor', '#ddeeff');
-
-        try {
-            $result = static::getContainer()->get(ScssPhpCompiler::class)->compileString(
-                new \Shopware\Storefront\Theme\CompilerConfiguration([]),
-                '$no-theme-custom-css-backgroundcolor: #aabbcc; $no-theme-custom-css-fontcolor: #ddeeff; ' . $testScss
-            );
-
-            static::assertStringContainsString('.test-selector-app', $result);
-            static::assertStringContainsString('#aabbcc', $result);
-            static::assertStringContainsString('#ddeeff', $result);
-        } finally {
-            $this->eventDispatcher->removeSubscriber($subscriber);
-        }
-    }
-
-    public function testCompilesPluginAndAppCssWithNullValueHandling(): void
-    {
-        $this->loadAppsFromDir(__DIR__ . '/fixtures/Apps/noThemeCustomCss');
-
-        $testScss = <<<'SCSS'
-.test-selector-plugin {
-    background: $simple-plugin-backgroundcolor;
-    color: $simple-plugin-fontcolor;
-    border: $simple-plugin-bordercolor;
-}
-.test-selector-app {
-    background: $no-theme-custom-css-backgroundcolor;
-    color: $no-theme-custom-css-fontcolor;
-    border: $no-theme-custom-css-bordercolor;
-}
-SCSS;
-
-        $configService = $this->getConfigurationService([
-            new SimplePlugin(true, __DIR__ . '/fixtures/SimplePlugin'),
-        ]);
-
-        $storefrontPluginRegistry = $this->getStorefrontPluginRegistry([
-            new SimplePlugin(true, __DIR__ . '/fixtures/SimplePlugin'),
-        ]);
-
-        $subscriber = new ThemeCompilerEnrichScssVarSubscriber($configService, $storefrontPluginRegistry);
-        $this->eventDispatcher->addSubscriber($subscriber);
-
-        $sysConfigService = static::getContainer()->get(SystemConfigService::class);
-        $sysConfigService->set('SimplePlugin.config.simplePluginBackgroundcolor', '#fff');
-        $sysConfigService->set('SwagNoThemeCustomCss.config.noThemeCustomCssBackGroundcolor', '#aaa');
-        // Note: bordercolor and fontcolor are intentionally NOT set to test null value handling
-
-        try {
-            // Build variables with nulls
-            $variables = '$simple-plugin-backgroundcolor: #fff; ';
-            $variables .= '$simple-plugin-fontcolor: #eee; ';
-            $variables .= '$simple-plugin-bordercolor: null; ';
-            $variables .= '$no-theme-custom-css-backgroundcolor: #aaa; ';
-            $variables .= '$no-theme-custom-css-fontcolor: #eee; ';
-            $variables .= '$no-theme-custom-css-bordercolor: null; ';
-
-            $result = static::getContainer()->get(ScssPhpCompiler::class)->compileString(
-                new \Shopware\Storefront\Theme\CompilerConfiguration([]),
-                $variables . $testScss
-            );
-
-            // Verify plugin selector with values
-            static::assertStringContainsString('.test-selector-plugin', $result);
-            static::assertStringContainsString('background:#fff', str_replace(' ', '', $result));
-            static::assertStringContainsString('color:#eee', str_replace(' ', '', $result));
-
-            // Verify app selector with values
-            static::assertStringContainsString('.test-selector-app', $result);
-            static::assertStringContainsString('background:#aaa', str_replace(' ', '', $result));
-
-            // IMPORTANT: Verify that border properties with null values are omitted
-            // SCSS omits property definitions when variables have null value
-            $normalizedResult = str_replace([' ', "\n", "\r"], '', strtolower($result));
-            static::assertStringNotContainsString(
-                'border:',
-                $normalizedResult,
-                'Border properties should be omitted when variable value is null'
-            );
-        } finally {
-            $this->eventDispatcher->removeSubscriber($subscriber);
-        }
     }
 
     // ===================================
@@ -430,10 +170,7 @@ SCSS;
             'FEATURE_NEXT_TEST_2' => ['default' => false],
         ]);
 
-        // Get the feature flag mixin
-        $featureMixin = file_get_contents(
-            __DIR__ . '/../../../../src/Storefront/Resources/app/storefront/src/scss/abstract/functions/feature.scss'
-        );
+        $featureMixin = '@function feature($feature-flag) { @return map-get($sw-features, $feature-flag); }';
 
         // Inject $sw-features variable (normally done by ThemeCompiler)
         $allFeatures = Feature::getAll();
@@ -461,7 +198,7 @@ SCSS;
 SCSS;
 
         $result = static::getContainer()->get(ScssPhpCompiler::class)->compileString(
-            new \Shopware\Storefront\Theme\CompilerConfiguration([]),
+            new CompilerConfiguration([]),
             $featureVariables . "\n" . $featureMixin . "\n" . $testScss
         );
 
@@ -508,7 +245,7 @@ SCSS;
         $vendorPath = __DIR__ . '/fixtures/ThemeWithScssVendorImports/Storefront/Resources/app/storefront/vendor';
 
         $result = static::getContainer()->get(ScssPhpCompiler::class)->compileString(
-            new \Shopware\Storefront\Theme\CompilerConfiguration([
+            new CompilerConfiguration([
                 'importPaths' => [
                     function (string $path) use ($vendorPath) {
                         if (str_starts_with($path, '~vendor/')) {
@@ -912,35 +649,6 @@ SCSS;
     }
 
     // ===================================
-    // Database Resilience Tests
-    // ===================================
-
-    public function testHandlesDatabaseException(): void
-    {
-        $configService = $this->getConfigurationServiceDbException([
-            new SimplePlugin(true, __DIR__ . '/fixtures/SimplePlugin'),
-        ]);
-
-        $storefrontPluginRegistry = $this->getStorefrontPluginRegistry([
-            new SimplePlugin(true, __DIR__ . '/fixtures/SimplePlugin'),
-        ]);
-
-        $event = new ThemeCompilerEnrichScssVariablesEvent(
-            [],
-            TestDefaults::SALES_CHANNEL,
-            Context::createDefaultContext()
-        );
-
-        $subscriber = new ThemeCompilerEnrichScssVarSubscriber($configService, $storefrontPluginRegistry);
-
-        // Should not throw exception
-        $subscriber->enrichExtensionVars($event);
-
-        // No variables should be added when a DB exception occurs
-        static::assertEmpty($event->getVariables());
-    }
-
-    // ===================================
     // End-to-End Compilation Tests
     // ===================================
 
@@ -1014,7 +722,7 @@ SCSS;
             'test'
         );
 
-        $componentCollection = new \Shopware\Storefront\Framework\Twig\Components\TwigComponentCollection([$mockComponent]);
+        $componentCollection = new TwigComponentCollection([$mockComponent]);
 
         // Create custom ThemeCompiler with mocked TwigComponentHelper
         $twigComponentHelper = $this->createMock(TwigComponentHelper::class);
@@ -1096,70 +804,5 @@ SCSS;
 
         $variablesContent = $this->tempFilesystem->read('theme-variables.scss');
         static::assertStringContainsString('$sw-asset-theme-url: \'http://localhost\'', $variablesContent);
-    }
-
-    // ===================================
-    // Helper Methods
-    // ===================================
-
-    /**
-     * @param array<int, Plugin> $plugins
-     */
-    private function getConfigurationService(array $plugins): ConfigurationService
-    {
-        return new ConfigurationService(
-            $plugins,
-            new ConfigReader(),
-            static::getContainer()->get(AppConfigReader::class),
-            static::getContainer()->get('app.repository'),
-            static::getContainer()->get(SystemConfigService::class),
-            static::getContainer()->get(LoggerInterface::class)
-        );
-    }
-
-    /**
-     * @param array<int, Plugin> $plugins
-     */
-    private function getConfigurationServiceDbException(array $plugins): ConfigurationService
-    {
-        return new ConfigurationServiceException(
-            $plugins,
-            new ConfigReader(),
-            static::getContainer()->get(AppConfigReader::class),
-            static::getContainer()->get('app.repository'),
-            static::getContainer()->get(SystemConfigService::class),
-            static::getContainer()->get(LoggerInterface::class)
-        );
-    }
-
-    /**
-     * @param array<int, Plugin> $plugins
-     */
-    private function getStorefrontPluginRegistry(array $plugins): StorefrontPluginRegistry
-    {
-        $kernel = $this->createMock(Kernel::class);
-        $kernel
-            ->method('getBundles')
-            ->willReturn($plugins);
-
-        return new StorefrontPluginRegistry(
-            $kernel,
-            static::getContainer()->get(StorefrontPluginConfigurationFactory::class),
-            static::getContainer()->get(ActiveAppsLoader::class)
-        );
-    }
-}
-
-/**
- * @internal
- */
-class ConfigurationServiceException extends ConfigurationService
-{
-    /**
-     * @throws Exception
-     */
-    public function checkConfiguration(string $domain, Context $context): bool
-    {
-        throw new \Doctrine\DBAL\Platforms\Exception\InvalidPlatformVersion('any');
     }
 }
